Only cleanup vendor in distclean when .git directory exists

+++ Makefile
@@ -59,7 +59,7 @@ distclean: clean
 	$(RM) ssh_host_ecdsa_key.pem ssh_host_ed25519_key.pem ssh_host_rsa_key.pem
 	$(RM) ./tags ./GPATH ./GRTAGS ./GTAGS
 	$(RM) -r ./log/
-	$(RM) -r ./vendor/
+	test -d ./.git && $(RM) -r ./vendor/ || :
 
 ##############################################################################
 # Target: tidy
